增加热门的缓存

+++ designer-mcap/src/main/java/com/bruce/designer/macp/api/command/hot/HotAlbumsCommand.java
@@ -62,7 +62,7 @@ public class HotAlbumsCommand extends AbstractApiCommand implements Initializing
     	int hostId = context.getUserId();
     	
     	String modeStr = context.getStringParams().get("mode");
-    	int mode = NumberUtils.toInt(modeStr, 0);
+    	short mode = NumberUtils.toShort(modeStr, (short)0);
     	
 		if(logger.isDebugEnabled()){
             logger.debug("MCS查询热门专辑，mode: "+mode);
