+++ designer-service/src/main/java/com/bruce/designer/service/IHotService.java
@@ -8,11 +8,11 @@ import com.bruce.designer.model.User;
 
 public interface IHotService {
 
-	public static final int HOURLY_FLAG = 0;
-	public static final int DAILY_FLAG = 1;
-	public static final int WEEKLY_FLAG = 2;
-	public static final int MONTHLY_FLAG = 3;
-	public static final int YEARLY_FLAG = 4;
+	public static final short HOURLY_FLAG = 0;
+	public static final short DAILY_FLAG = 1;
+	public static final short WEEKLY_FLAG = 2;
+	public static final short MONTHLY_FLAG = 3;
+	public static final short YEARLY_FLAG = 4;
 
 	// 获取热门tag
 	public List<Tag> getHotTags(int limit);
@@ -21,10 +21,10 @@ public interface IHotService {
 	public List<Tag> calcHotTags(int limit);
 
 	// 获取热门作品
-	public List<Album> fallLoadHotAlbums(int mode, int limit);
+	public List<Album> fallLoadHotAlbums(short mode, int limit);
 
 	// 获取热门设计师
-	public List<User> fallLoadHotDesigners(int mode, int limit);
+	public List<User> fallLoadHotDesigners(short mode, int limit);
 
 	// public List<User> fallLoadHotDesigners(int start, int limit);
